verdin-imx8mm: port 7-inch LVDS panel overlay from i.MX8MP
Allows using the display with the Verdin i.MX8M Mini module. device tree
based on our imx8mp dts and Toradex imx8mm examples.

I imported verdin-imx8_mipi-dsi-to-sn65dsi84.dtsi from
https://git.toradex.com/cgit/device-tree-overlays.git
because I didn't want to duplicate the source code for that file.
Meanwhile, it cannot be easily imported in our Yocto recipe because
the device-tree-overlays Toradex recipe doesn't export its sources.
Since this file is under the same license as the rest of our repository
(GPL-v2.0), it is acceptable to duplicate the file here.

Change-Id: I5305e9de9442b6096d4e6747695cc8b17aa0d84e