colibri-imx8x: port 7inch overlay from imx8mp
Adapt the verdin-imx8mp dts for the colibri SoM based on Toradex
examples.

This must be used along with the non-default base DTB for the IRIS v2
board. There's no LVDS connector on the other Colibri eval boards and
we need the 3v3 supply from the Iris v2 to be pinmuxed.

Change-Id: I48c12ca8514484f802817359bba6faf8107f54a9