verdin-imx95: add common verdin-imx95 and 10 inches display support
Add common optologic support for verdin-imx95 SOM as well as specific
configuration for the 10inch display.
This support is based on the device-tree overlay for the Toradex
Capacitive Touch Display 10.1" LVDS [1], as well as on other optologic
device-trees.
The ldb clocks frequencies are tuned to match the typical frequency
of the screen.

[1]: https://git.toradex.com/device-tree-overlays.git/tree/overlays/verdin-imx95_panel-cap-touch-10inch-lvds_overlay.dts?h=toradex_6.12-2.0.x-imx

Change-Id: I1fe90f79a651ee1696e2d451619ff84a6c2a4207